feat : Line -> Sections로 로직 분리/이동
